Refactor account state db
